Aha, this may be the fix.
htmx.ajax is *asynchronous*, so I should wait for the first call to finish before starting the second one.

Not sure why that matters, but I do recall seeing exceptions in the browser console when I had the bot sending messages
very frequently -- presumably that caused the browser to start lots of ajaxes concurrently, which presumably causes
chaos and/or tsuris.