WIP -- fixing fixing fixing
pretty sure I'm confusing myself by conflating two meanings of the word "hand"

This django app defines that as basically "one board of 52 cards plus four players who will play, or are playing, or
have played it"; but the library defines it as "up to 13 cards of a given suit held by one player".

Gonna whup the library to rename that to "holding".