Try to ensure that HOST_HOSTNAME is never empty in docker-compose
I saw an exception report from the bot running on my laptop, which Sentry tagged as "production":

https://eric-hanchrow.sentry.io/issues/6028821664/?project=4507936354205696

I assume this is because there's no environment variable named HOST_HOSTNAME in the bot's container, which is equivalent
to the empty string, which does not startswith("Erics-Work-MacBook-Pro"), so it figures it's production.

And I assume *that* happened because the compose file element for the "bot" service defines its own ENVIRONMENT, which I
guess totally overwrites that from the "django" service, which I'd hoped it would "inherit".