Rethink "Player.currently_seated"
The old way:

currently_seated was a boolean field that means what it says.  But if it's True, and you want to know *which* seat the
player is at, you have to examine all the Seats that refer to this player, and somehow pick the one that is
"current". We did that by choosing the one whose primary key is greatest, but that seems pretty iffy.

The new way:

currently_seated is a property that simply asks if the (new) field "current_seat" is None.  That field, of course, is a
ForeignKey to the relevant Seat object.