Fix stupid bug in that method.
That pointed out that "test_running_tournament_irrelevant_players" wasn't quite what I wanted.

I've decided that irrelevant players should not be allowed to see anything.  It's true that they'll never play that board;
but it's just too easy for a player to create an incognito window and either look at the hand while not logged in, or
else sign up with a "burner" account in order to peek.

(Also: Rewrite hand_at_which_we_played_board to use a single giant query)