Include a simple serial number with each hand event
The dream is: the bot will only fetch the complete hand once, when it starts; thenceforth it'll just ingest the new-call
and new-play events, and will keep its transcript up to date that way; and will compare these serial numbers against the
count of events it's received.  If the counts differ, it'll restart.

In fact, I've got it doing that, but ... it restarts too often :-|