Failed attempt to repro the problem.
I ran

    just restore < bot-thinks-its-bobs-turn-2025-07-08T18:39:37+0000.sql && just manage repro

expecting to see a difference between the two sets of "prints", but they are the same, and both look correct: they agree
with the web page, and show that "_marlo-polk" is the next caller.

And yet when I actually run the bot (instead of "just manage repro") it sits there telling me that "bob is human", and
doing nothing else.